requirements.txt
altair==5.4.1
google-analytics-data==0.18.12
google-auth==2.35.0
google-auth-oauthlib==1.2.1
numpy==2.1.2
pandas==2.2.3
pillow==10.4.0
psycopg2-binary==2.9.9
requests==2.32.3
SQLAlchemy==2.0.35
streamlit==1.39.0

altair==5.4.1
attrs==24.2.0
blinker==1.8.2
briefer==0.0.13
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docker==7.1.0
gitdb==4.0.11
GitPython==3.1.43
google-analytics-data==0.18.12
google-api-core==2.20.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
narwhals==1.9.1
numpy==2.1.2
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==307
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.35
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
